+++ src/modules/labels/client.ts
@@ -2,5 +2,5 @@
 
 export { useCreateLabel, useUpdateLabel } from './client/query/mutations'
 export { useLabels } from './client/query/queries'
-export { labelKeys } from './cache'
+export { labelKeys } from './query-cache'
 export type { CreateLabel, Label, UpdateLabel } from './domain/label'